[libc++] Use proper shell escaping in the executors
This was originally committed as 8193dfeaa3b0 and reverted in 081b4ac6c6ba.
The issue was that shell builtins were being escaped too, and apparently
Bash won't execute a builtin when it is quoted e.g. '!'. Instead, it
thinks it's a command and it can't find it.

Re-committing the change with that issue fixed.