[libc++] Use proper shell escaping in the executors
This was originally committed as fdadcee7c593 and reverted in dd56db09efa3.
The issue was that shell builtins were being escaped too, and apparently
Bash won't execute a builtin when it is quoted e.g. '!'. Instead, it
thinks it's a command and it can't find it.

Re-committing the change with that issue fixed.