[libc++] Re-enable warnings in the new format
When the new libc++ test format was enabled, warnings were accidentally
dropped cause they were not part of the %{compile_flags} substitution.
This commit adds them back, however `-Werror` is only used for non-verify
tests (cause it doesn't make sense for verify tests).

This commit is a re-application of 47a2ad0ecdcd, which was reverted in
a42dde9cdae4 because it broke the C++03 bot. This failure should have
been fixed in be1e9be97379.