[libc++] Re-enable warnings in the new format
When the new libc++ test format was enabled, warnings were accidentally
dropped cause they were not part of the %{compile_flags} substitution.
This commit adds them back, however `-Werror` is only used for non-verify
tests (cause it doesn't make sense for verify tests).

This commit is a re-application of 9c1d9eb8960d, which was reverted in
9d1c13e998fa because it broke the C++03 bot. This failure should have
been fixed in d9739234b956.