Revert switch based variant temporarily.
There are currently some failures caused by this change internally. I'm working
to debug them and hopefully these series of patches should be recommitted by
the end of the week.

Thank you to Micheal Park for the contributions, and for allowing the temporary
rollback.

The commits reverted by this change are:

91272898ae1ba7b78c4d37b3029275083ab2c9d5
230d844af523f95a2e85d8ee5652d5b129d4e44d
87e0eeaba8d996385eb6229a2283a73afdef84e1
3a58de0190b1eb7c3edde8a671f8e37b55efa721