[runtimes] Revert the libc++ __config_site change
This is a massive revert of the following commits (from most revent to oldest):

	2b9b7b5775a1d8fcd7aa5abaa8fc0bc303434f1a.
	529ac33197f6408952ae995075ac5e2dc5287e81
	df1942edca1b50dd50fa5042ac88e2b01abc8dc9
	a6a2f409f090b6fff1b872f8974bd1578e11b983
	350367771ad61af7fecb831cadde1e1cc3e96439
	eaaf9c2c148ea9c41803bfc2d759052901828bba

After checking-in the __config_site change, a lot of things started breaking
due to widespread reliance on various aspects of libc++'s build, notably the
fact that we can include the headers from the source tree, but also reliance
on various "internal" CMake variables used by the runtimes build and compiler-rt.

These were unintended consequences of the change, and after two days, we
still haven't restored all the bots to being green. Instead, now that I
understand what specific areas this will blow up in, I should be able to
chop up the patch into smaller ones that are easier to digest.

See https://reviews.llvm.org/D89041 for more details on this adventure.